+++ content/blog/dojo-kun.md
@@ -5,73 +5,75 @@ layout: post
 author:
   name: Salrion
 name: null
-title: "Life: Dojo Kun - 道場訓 - Preceptos del Dojo"
+title: 'Life: Dojo Kun - 道場訓 - Preceptos del Dojo'
 tags:
   - life
 id: dojokun
-year: 2017-05-07
+year: 2017
+date: 2017-05-07
 description: Dojo kun, reglas dojo para practicantes de Karate Shotokan. Reglas
   básicas que todo practicante debe tratar de cumplir, tanto dentro como fuera
   del dojo.
 ---
+
 **一丶 人格完成に努むること。**
-*Hitotsu, Jinkaku Kansei Ni Tsutomuru Koto.* <br>
+_Hitotsu, Jinkaku Kansei Ni Tsutomuru Koto._ <br>
 Lo primero, esforzarse por la perfección del carácter. <br>
 Buscar la perfección del carácter <br>
 
-| Vocabulario | |
-|-------------|--|
-| Jinkaku | Personalidad |
-| Kansei  | Completar, finalizar |
-| Tsutomuru | Trabajo, esfuerzo |
+| Vocabulario |                      |
+| ----------- | -------------------- |
+| Jinkaku     | Personalidad         |
+| Kansei      | Completar, finalizar |
+| Tsutomuru   | Trabajo, esfuerzo    |
 
 <br>
 
 **一丶 誠の道を守ること。**
-*Hitotsu, MaKoto No Michi O Mamoru Koto.*<br>
+_Hitotsu, MaKoto No Michi O Mamoru Koto._<br>
 Lo primero, defender los caminos de la verdad.<br>
 Ser leal y sincero<br>
 
-| Vocabulario | |
-|-------------|--|
-| **Makoto** | Verdad |
-| **Michi**  | Carretera o camino |
-| **Mamoru** | Proteger |
+| Vocabulario |                    |
+| ----------- | ------------------ |
+| **Makoto**  | Verdad             |
+| **Michi**   | Carretera o camino |
+| **Mamoru**  | Proteger           |
 
 <br>
 
 **一丶努力の精神を養うこと。**
-*Hitotsu, Doryoku No Seichin O Yashinau Koto.*<br>
+_Hitotsu, Doryoku No Seichin O Yashinau Koto._<br>
 Lo primero, fomentar el espíritu del esfuerzo.<br>
 Poner el máximo esfuerzo y constancia<br>
 
-| Vocabulario | |
-|------------- |--|
-| **Doryoku**  | Esfuerzo |
-| **Seichin**  | Espiritu |
+| Vocabulario  |                     |
+| ------------ | ------------------- |
+| **Doryoku**  | Esfuerzo            |
+| **Seichin**  | Espiritu            |
 | **Yashinau** | Cultivar o fomentar |
 
 <br>
 
 **一丶礼儀を重んずること。**
-*Hitotsu, Reigi O Omonzuru Koto.*<br>
+_Hitotsu, Reigi O Omonzuru Koto._<br>
 Lo primero, honrar los principios de etiqueta.<br>
 Respeto a los demás<br>
 
-| Vocabulario | |
-|-------------|--|
+| Vocabulario  |                     |
+| ------------ | ------------------- |
 | **Reigi**    | Cortesia o etiqueta |
-| **Omonzuru** | Honor, Honrar |
+| **Omonzuru** | Honor, Honrar       |
 
 <br>
 
 **一丶血気の勇を戒むること。**
-*Hitotsu, Kekki No Yu O Imashimuru Koto.*<br>
+_Hitotsu, Kekki No Yu O Imashimuru Koto._<br>
 Lo primero, guardarse contra el valor impetuoso.<br>
 Autocontrol, reprimir la violencia<br>
 
-| Vocabulario | |
-|-------------|--|
+| Vocabulario     |                 |
+| --------------- | --------------- |
 | **Kekki**       | Vigor, ardiente |
 | **Kekki no Yu** | sangre ardiente |
-| **Imashimuru**  | Dejarse llevar |
+| **Imashimuru**  | Dejarse llevar  |
