+++ content/blog/flexget.md
@@ -1,7 +1,8 @@
 ---
 name: 'flexget-instalacion-y-configuracion-raspberry-pi'
 title: 'Flexget: Instalación y configuración en Raspberry'
-year: 2016-11-06
+year: 2016
+date: 2016-11-06
 id: 'flexget-instalacion-y-configuracion-raspberry-pi'
 description: flexget, instalación y configuración en Raspberry pi para sacar provecho a la automatización de diversas tareas, como uso de Transmision y subtitulos
 layout: post
