+++ content/blog/instaling-transmision.md
@@ -1,7 +1,8 @@
 ---
 name: 'instalar-transmision'
 title: 'Transmision: Instalación en Linux'
-year: 2016-06-11
+year: 2016
+date: 2016-06-11
 id: 'instalar-transmision-linux'
 description: Notas de instalación de transmision en Linux, herramienta de descarga de torrents, tanto en formato cli, como web y configuración base
 layout: post
