+++ nuxt.config.ts
@@ -1,5 +1,3 @@
-//import { defineNuxtConfig } from 'nuxt';
-
 // https://v3.nuxtjs.org/api/configuration/nuxt.config
 export default defineNuxtConfig({
   modules: ['@nuxt/content'],
