+++ package.json
@@ -7,7 +7,7 @@
     "preview": "nuxt preview"
   },
   "devDependencies": {
-    "eslint": "^8.21.0",
+    "eslint": "8.22.0",
     "@nuxt/content": "^2.1.0",
     "nuxt": "^3.0.0",
     "prettier": "^2.7.1",
