--- package.json.old
@@ -1,57 +1,0 @@
-{
-  "name": "salrionblog",
-  "version": "1.0.0",
-  "description": "Blog personal, apuntes sobre tecnología, linux, programación, tecnologías, forma de vida. Una pequeño repositorio de cosas de interes.",
-  "author": "salrio",
-  "private": true,
-  "scripts": {
-    "build": "nuxt build",
-    "dev": "nuxt dev",
-    "generate": "nuxt generate",
-    "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
-    "test": "jest"
-  },
-  "lint-staged": {
-    "*.{js,vue}": "npm run lint",
-    "*.{css,vue}": "stylelint"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "dependencies": {
-    "@nuxt/content": "^1.14.0",
-    "@nuxtjs/sitemap": "^2.4.0",
-    "netlify-cms-app": "^2.15.19",
-    "nuxt": "^2.15.6",
-    "nuxt-trailingslash-module": "^1.2.0",
-    "vue-cookieconsent-component": "^1.2.0",
-    "vue-gtag": "^1.16.1",
-    "vue-if-bot": "^1.2.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/eslint-parser": "^7.14.3",
-    "@babel/eslint-plugin": "^7.13.16",
-    "@nuxtjs/eslint-config": "^6.0.0",
-    "@nuxtjs/eslint-module": "^3.0.2",
-    "@nuxtjs/google-analytics": "^2.4.0",
-    "@nuxtjs/stylelint-module": "^3.1.0",
-    "eslint": "^7.27.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-nuxt": "^2.0.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-vue": "^7.9.0",
-    "fibers": "^5.0.0",
-    "husky": "^4.3.8",
-    "jest": "^24.1.0",
-    "lint-staged": "^10.5.4",
-    "prettier": "^1.19.1",
-    "sass": "^1.34.0",
-    "sass-loader": "^10.2.0",
-    "stylelint": "^10.1.0",
-    "typescript": "^4.2.4",
-    "webpack": "^4.36.0"
-  }
-}
